add recovery cli and split up endpoint
this will be useful for wasm later but need it now for recovery split
