+++ packages/xreal_one_driver/src/proto/usb/pilot_update.rs
@@ -54,20 +54,25 @@ impl UsbDevice {
     ) -> Result<(), anyhow::Error> {
         let header = PilotFirmwareHeader::load(update)?;
 
-        self.send_message::<PilotUpdateStart>(header).await?;
+        self.endpoint
+            .send_message::<PilotUpdateStart>(header)
+            .await?;
         progress.transmit(PilotFirmwareHeader::LEN);
 
         let mut position = PilotFirmwareHeader::LEN;
         while position < update.len() {
             let end_position = std::cmp::min(position + 1002, update.len());
-            self.send_message::<PilotUpdateTransmit>(RawRequest(&update[position..end_position]))
+            self.endpoint
+                .send_message::<PilotUpdateTransmit>(RawRequest(&update[position..end_position]))
                 .await?;
             progress.transmit(end_position - position);
 
             position = end_position;
         }
 
-        self.send_message::<PilotUpdateFinish>(Empty).await?;
+        self.endpoint
+            .send_message::<PilotUpdateFinish>(Empty)
+            .await?;
 
         Ok(())
     }
